+++ libs/newsletters-data-client/src/index.ts
@@ -5,7 +5,7 @@ export * from './lib/api-response-type';
 export * from './lib/wizard-button-type';
 export * from './lib/testQuestionaireSchema';
 export * from './lib/emailEmbedSchema';
-export * from './lib/storage';
+export * from './lib/draft-storage';
 export * from './lib/transformDataToLegacyNewsletter';
 export * from './lib/schema-helpers';
 export * from './lib/transformWizardData';
