+++ libs/newsletters-data-client/src/lib/draft-storage/InMemoryDraftStorage.ts
