+++ libs/newsletters-data-client/src/lib/draft-storage/S3DraftStorage/objectToDraftWithId.ts
