+++ libs/newsletters-data-client/src/lib/draft-storage/S3DraftStorage/s3Functions.ts
